+++ changelog.md
@@ -21,6 +21,7 @@ __[maintenance]__
 * Update certificate information for MariaDB 11.4 installer.
 * Update certificate information for MariaDB 11.8 installer.
 * Update certificate information for Node.js installer.
+* Update certificate information for Opera installer.
 * Update certificate information for TeamViewer installer.
 
 ## Version 2026.05.03.0